Refactor access token
